dddd

+++ ft_printf.c
@@ -27,7 +27,7 @@ static int	ft_print_str(va_list args, const char print_str)
 	else if (print_str == 'u')
 		len += ft_putunbr(va_arg(args, unsigned int));
 	else if (print_str == 'x' || print_str == 'X')
-		len += ft_puthex(va_arg(args, unsigned long), print_str);
+		len += ft_puthex(va_arg(args, unsigned long long), print_str);
 	else if (print_str == 'p')
 		len += ft_putptr(va_arg(args, unsigned long *));
 	else if (print_str == '%')
